refactor(services/fs): extract implementation to core
