+++ core/src/services/fs/error.rs
@@ -15,23 +15,17 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#[cfg(feature = "services-fs")]
-mod core;
-#[cfg(feature = "services-fs")]
-mod delete;
-#[cfg(feature = "services-fs")]
-mod error;
-#[cfg(feature = "services-fs")]
-mod lister;
-#[cfg(feature = "services-fs")]
-mod reader;
-#[cfg(feature = "services-fs")]
-mod writer;
+use crate::raw::*;
+use crate::*;
 
-#[cfg(feature = "services-fs")]
-mod backend;
-#[cfg(feature = "services-fs")]
-pub use backend::FsBuilder as Fs;
-
-mod config;
-pub use config::FsConfig;
+/// Parse error response into Error.
+pub(super) fn parse_error(e: std::io::Error) -> Error {
+    match e.kind() {
+        std::io::ErrorKind::AlreadyExists => Error::new(
+            ErrorKind::ConditionNotMatch,
+            "The file already exists in the filesystem",
+        )
+        .set_source(e),
+        _ => new_std_io_error(e),
+    }
+}
